+++ Tests/BugsnagMetadataRedactionTest.m
@@ -1,0 +1,102 @@
+//
+//  BugsnagMetadataRedactionTest.m
+//  Tests
+//
+//  Created by Jamie Lynch on 15/04/2020.
+//  Copyright © 2020 Bugsnag. All rights reserved.
+//
+
+#import <XCTest/XCTest.h>
+#import "BugsnagEvent.h"
+
+@interface BugsnagEvent ()
+- (NSDictionary *)toJson;
+
+@property NSArray *redactedKeys;
+@end
+
+@interface BugsnagMetadataRedactionTest : XCTestCase
+
+@end
+
+@implementation BugsnagMetadataRedactionTest
+
+- (void)testEmptyRedaction {
+    BugsnagEvent *event = [self generateEventWithMetadata:@{
+            @"password": @"hunter2",
+            @"some_key": @"2fa0"
+    }];
+
+    NSDictionary *payload = [event toJson];
+    NSDictionary *section = payload[@"metaData"][@"custom"];
+    XCTAssertNotNil(section);
+    XCTAssertEqualObjects(@"hunter2", section[@"password"]);
+    XCTAssertEqualObjects(@"2fa0", section[@"some_key"]);
+}
+
+- (void)testDefaultRedaction {
+    BugsnagEvent *event = [self generateEventWithMetadata:@{
+            @"password": @"hunter2",
+            @"some_key": @"2fa0"
+    }];
+    event.redactedKeys = @[@"password"];
+
+    NSDictionary *payload = [event toJson];
+    NSDictionary *section = payload[@"metaData"][@"custom"];
+    XCTAssertNotNil(section);
+    XCTAssertEqualObjects(@"[REDACTED]", section[@"password"]);
+    XCTAssertEqualObjects(@"2fa0", section[@"some_key"]);
+}
+
+- (void)testNestedRedaction {
+    BugsnagEvent *event = [self generateEventWithMetadata:@{
+            @"user_auth": @{
+                    @"password": @"123456",
+                    @"authority": @"admin",
+                    @"meta": @{
+                            @"password": @"fff455"
+                    }
+            },
+            @"some_key": @"2fa0"
+    }];
+    event.redactedKeys = @[@"password"];
+
+    NSDictionary *payload = [event toJson];
+    NSDictionary *section = payload[@"metaData"][@"custom"];
+    XCTAssertNotNil(section);
+    XCTAssertEqualObjects(@"[REDACTED]", section[@"user_auth"][@"meta"][@"password"]);
+    XCTAssertEqualObjects(@"[REDACTED]", section[@"user_auth"][@"password"]);
+    XCTAssertEqualObjects(@"admin", section[@"user_auth"][@"authority"]);
+    XCTAssertEqualObjects(@"2fa0", section[@"some_key"]);
+}
+
+- (void)testNonDefaultKeys {
+    BugsnagEvent *event = [self generateEventWithMetadata:@{
+            @"user_auth": @{
+                    @"password": @"123456",
+                    @"authority": @"admin"
+            },
+            @"some_key": @"2fa0",
+            @"foo": @"gasdf"
+    }];
+    event.redactedKeys = @[@"authority", @"some_key"];
+
+    NSDictionary *payload = [event toJson];
+    NSDictionary *section = payload[@"metaData"][@"custom"];
+    XCTAssertNotNil(section);
+    XCTAssertEqualObjects(@"123456", section[@"user_auth"][@"password"]);
+    XCTAssertEqualObjects(@"[REDACTED]", section[@"user_auth"][@"authority"]);
+    XCTAssertEqualObjects(@"[REDACTED]", section[@"some_key"]);
+}
+
+- (BugsnagEvent *)generateEventWithMetadata:(NSDictionary *)data {
+    return [[BugsnagEvent alloc] initWithKSReport:@{
+            @"user": @{
+                    @"metaData": @{
+                            @"custom": data
+                    }
+            }
+    }];
+}
+
+@end